feat: 스크랩 조회(#9)
